Crate oscore

An experimental OSCORE implementation with EDHOC key exchange, intended for embedded devices.

The EDHOC implementation is based on draft 14 of the EDHOC specification. It only does authentication with raw public keys (RPK), so it covers the asymmetric authentication scenario, but not the symmetric one using pre-shared keys (PSK). On the OSCORE side, it does key derivation using the master secret and master salt, which can be established with EDHOC.

There is a demo implementation using this library, with a resource server on an STM32F3, a client on an STM32F4 and a CoAP proxy running on a Raspberry Pi.

Security

This should not currently be used in production code, use at your own risk.

Modules

edhoc
Ephemeral Diffie-Hellman Over COSE (EDHOC) to establish an OSCORE context.

oscore
OSCORE implementation deriving keys from a master secret and master salt.